Worker methods for talking to REST-ful API.
This provides 'restGet', 'restDestroy', and 'restSave' as methods in the "data" component
of the Conduit Worker.  Only 'restGet' is used, but in the future this will allow all
server interaction to happen on the worker instead of saving/destroying on the UI
thread.

As a side effect, this also addresses Issue #31.